Update dogm-graphic.h
The command in line 444 should be the same as in line 441 and according to EA datasheet and UC1701 datasheet the last four (not 5) bits are the page-address. 
It probably only rarely becomes an issue, if users never target an address > 15
The command for setting temperature compensation mode is [TC, 0, 0, 1], so it should be 0x90 for high and 0x10 for low (also stated in both datasheets, so I assume correct).

Did not have the chance to confirm on hardware yet, whether the mistake is in the code or the datasheets, but at least the first is either incorrect on line 444 or 441. Thought I better post it now, before I forget.